backend/boundary/API/requirements.txt
django
djangorestframework
environs
requests
djangorestframework-simplejwt